modelcontactsurface working with pressure

+++ mecacell/basicworld.hpp
@@ -35,7 +35,7 @@ class BasicWorld {
 
  protected:
 	float_t dt = 1.0 / 100.0;  // interval btwn updates
-	int nbAddedCells = 0;      // used for each added cell's unique id
+	size_t nbAddedCells = 0;   // used for each added cell's unique id
 
 	// space partition hashmap for cells
 	SpacePartition<Cell *> cellSpacePartition =
